fix: (part 1?) only auto-apply in the end if it's possible w/o exploding (regression introduced in 0657970dd2303f264036606e0cdc06b6854a0104)
currently `canAndShouldBeApplying` is only in that part of the logic.

to be fair, the problem arises only when we try to apply
before the rebase was finished.

but, it very well could be that the user (even tho unlikely) that user
runs `git-stacked-rebase` while a regular `git-rebase` is in progress,
thus causing the same issue.
TODO so we need to handle this everywhere, and the approach might be
different as well.

Signed-off-by: Kipras Melnikovas <kipras@kipras.org>