refactor: extract rewritten-list reading logic into apply.ts
...because completely related to the logic of "doesNeedToApply",
"markThatNeedsToApply", "markThatApplied", etc.

thought about extracting the above utils into a separate, "rewrittenList.ts" file,
but it doesn't make sense to separate the utils out from "apply" itself,

because if we have the "rewrittenList.ts" file, then it'd sound
reasonable to move the "reduce-path"'s "combineRewrittenLists" there as well,
but it doesnt, because the apply logic is not related to the
parsing/combinning logic.

Signed-off-by: Kipras Melnikovas <kipras@kipras.org>