use adaptive branch boundary finder behavior as the default in `branchSequencer`
realised that `--branch-sequencer --exec` should be adaptive as well,
just like forcePush,
because it can be ran anytime -- both if stacked rebase is not in
progress (needs an apply), and when it's done.

another option would be to allow the user to specify the wanted
behavior, so that it'd be explicit & we'd fail if the expectations
aren't met,
but i haven't seen a use-case for this yet,
and either way, the current solution should be good in any case,
while the 'expectations' idea can be added later if needed.

Signed-off-by: Kipras Melnikovas <kipras@kipras.org>